+++ Platform/TickZoomAPI1.0/Engine/ProviderSimulator.cs
@@ -30,8 +30,8 @@ namespace TickZoom.Api
 {
 	public interface ProviderSimulator
 	{
-		void PositionChange(Tick tick, double signal, ActiveList<LogicalOrder> inputOrders);
+		void PositionChange(Tick tick, int signal, ActiveList<LogicalOrder> inputOrders);
 		void ProcessOrders(Tick tick);
-		double ActualPosition {	get; }
+		int ActualPosition {	get; }
 	}
 }
